requirements.txt
aio-pika==9.5.6
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiormq==6.8.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
distro==1.9.0
dnspython==2.7.0
email-reply-parser==0.5.12
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
frozenlist==1.8.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langsmith==0.4.21
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.4
oauthlib==3.3.1
openai==1.102.0
orjson==3.11.3
packaging==25.0
pamqp==3.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pyparsing==3.2.5
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.35.1
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.43
starlette==0.47.3
structlog==25.4.0
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typer==0.17.3
typing-extensions==4.15.0
typing-inspection==0.4.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
